Fsync the affected files/directories only in pg_rewind.
Previously in the end, pg_rewind would fsync the whole pgdata directory on the
target, but that is a waste since usually just part of the files/directories on
the target are modified. Other files on the target should have been flushed
since pg_rewind requires a clean shutdown before doing the real work. This
would help the scenario that the target postgres instance includes millions of
files, which has been seen in a real environment.